Tools/utils/requirements.txt
python-dotenv
pika==1.3.2
pillow==11.0.0
pytz>=2023.3
